NW4R and RVL import from ogws, decompile m2d (#68)
* Import nw4r and revolution headers from ogws
* Fix existing code to work with new headers
* m2d almost matching
* Fix decomp context creation
* Match and link m2d
* Apply code review

+++ include/game/mLib/m_sphere.hpp
@@ -2,12 +2,10 @@
 
 #include <game/mLib/m_vec.hpp>
 #include <lib/egg/geom/eggSphere.h>
-#include <lib/MSL_C/float.h>
-#include <lib/MSL_C/math.h>
 
 class mSphere_c : public EGG::Sphere3f {
 public:
     mSphere_c() : EGG::Sphere3f(mVec3_c::Zero, 0.0f) {}
 
-    bool isZero() { return fabsf(mRadius) <= FLT_EPSILON; }
+    bool isZero() { return std::fabs(mRadius) <= FLT_EPSILON; }
 };
